+++ components/datetime/tests/patterns/structs.rs
